Fixes for comments from PR process
