+++ src/exceptions/cube-error-exception.ts
@@ -1,13 +1,4 @@
-export enum CubeValidationType {
-  FactTable = 'factTable',
-  FactTableColumnMissing = 'factTableColumnMissing',
-  DuplicateFact = 'duplicateFact',
-  Dimension = 'dimension',
-  DimensionNonMatchedRows = 'dimension_non_matched_rows',
-  Measure = 'measure',
-  NoFirstRevision = 'no_first_revision',
-  CubeCreationFailed = 'cube_creation_failed'
-}
+import { CubeValidationType } from '../enums/cube-validation-type';
 
 export class CubeValidationException extends Error {
   public type: CubeValidationType;
