+++ src/services/reference-data-handler.ts
@@ -16,7 +16,7 @@ import {
   loadCorrectReferenceDataIntoReferenceDataTable,
   loadReferenceDataIntoCube
 } from './cube-handler';
-import { createEmptyCubeWithFactTable } from '../utils/create-facttable';
+import { createEmptyCubeWithFactTable } from '../utils/create-fact-table';
 
 const sampleSize = 5;
